fix: mv system instruction to own pkg, rm use of aliases
